+++ packages/iov-bcp/src/transactions.ts
@@ -1,7 +1,7 @@
 import { ReadonlyDate } from "readonly-date";
 import { As } from "type-tagger";
 
-import { Preimage } from "./atomicswaptypes";
+import { Hash, Preimage } from "./atomicswaptypes";
 
 export enum Algorithm {
   Ed25519 = "ed25519",
@@ -266,7 +266,7 @@ export interface SwapOfferTransaction extends UnsignedTransaction {
    *
    * This is a SHA256 hash until we have a way to specifiy the hashing algorithm.
    */
-  readonly hash: Uint8Array;
+  readonly hash: Hash;
   readonly memo?: string;
 }