+++ packages/iov-ethereum/src/ethereumconnection.ts
@@ -22,6 +22,7 @@ import {
   ConfirmedTransaction,
   FailedTransaction,
   Fee,
+  Hash,
   isAtomicSwapIdQuery,
   isPubkeyQuery,
   Nonce,
@@ -767,7 +768,7 @@ export class EthereumConnection implements BcpAtomicSwapConnection {
       const recipient = toChecksummedAddress(
         Abi.decodeAddress(resultArray.slice(recipientBegin, recipientEnd)),
       );
-      const hash = resultArray.slice(hashBegin, hashEnd);
+      const hash = resultArray.slice(hashBegin, hashEnd) as Hash;
       const timeoutHeight = new BN(resultArray.slice(timeoutBegin, timeoutEnd)).toNumber();
       const amount = {
         quantity: new BN(resultArray.slice(amountBegin, amountEnd)).toString(),